+++ dashboard/src/routes/_dashboard/hosts/$hostId/delete.tsx
@@ -1,56 +1,28 @@
 import {
     createFileRoute,
     useNavigate,
-    defer,
-    Await
 } from "@tanstack/react-router";
 import {
     HostsDeleteConfirmationDialog,
-    fetchHost,
+    useRouterHostContext,
 } from "@marzneshin/features/hosts";
-import { Suspense } from "react";
-import {
-    AlertDialog,
-    AlertDialogContent,
-    Loading
-} from "@marzneshin/components";
 import { useDialog } from "@marzneshin/hooks";
 
 const HostDelete = () => {
     const [deleteDialogOpen, setDeleteDialogOpen] = useDialog(true);
-    const { host } = Route.useLoaderData()
+    const value = useRouterHostContext()
     const navigate = useNavigate({ from: "/hosts/$hostId/delete" });
 
-    return (
-        <Suspense fallback={<Loading />}>
-            <Await promise={host}>
-                {(host) => (
-                    <HostsDeleteConfirmationDialog
-                        open={deleteDialogOpen}
-                        onOpenChange={setDeleteDialogOpen}
-                        entity={host}
-                        onClose={() => navigate({ to: "/hosts" })}
-                    />
-                )}
-            </Await>
-        </Suspense>
+    return value && (
+        <HostsDeleteConfirmationDialog
+            open={deleteDialogOpen}
+            onOpenChange={setDeleteDialogOpen}
+            entity={value.host}
+            onClose={() => navigate({ to: "/hosts" })}
+        />
     );
 }
 
 export const Route = createFileRoute('/_dashboard/hosts/$hostId/delete')({
-    loader: async ({ params }) => {
-        const hostPromise = fetchHost({
-            queryKey: ["hosts", Number.parseInt(params.hostId)]
-        });
-
-        return {
-            host: defer(hostPromise)
-        }
-    },
-    errorComponent: () => (
-        <AlertDialog open={true}>
-            <AlertDialogContent>Host not found</AlertDialogContent>
-        </AlertDialog>
-    ),
     component: HostDelete,
 })
